Vector: Refacoring and code improvements
