+++ internal/generator/vector/conf_test.go
@@ -549,344 +549,6 @@ address = "0.0.0.0:24231"
 default_namespace = "collector"
 `,
 		}),
-		Entry("with multiple pipelines for elastic-search", generator.ConfGenerateTest{
-			CLSpec: logging.ClusterLoggingSpec{
-				Forwarder: &logging.ForwarderSpec{},
-			},
-			CLFSpec: logging.ClusterLogForwarderSpec{
-				Pipelines: []logging.PipelineSpec{
-					{
-						InputRefs: []string{
-							logging.InputNameApplication,
-							logging.InputNameInfrastructure,
-						},
-						OutputRefs: []string{"es-1"},
-						Name:       "pipeline1",
-					},
-					{
-						InputRefs: []string{
-							logging.InputNameAudit},
-						OutputRefs: []string{"es-1", "es-2"},
-						Name:       "pipeline2",
-					},
-				},
-				Outputs: []logging.OutputSpec{
-					{
-						Type: logging.OutputTypeElasticsearch,
-						Name: "es-1",
-						URL:  "https://es-1.svc.messaging.cluster.local:9200",
-						Secret: &logging.OutputSecretSpec{
-							Name: "es-1",
-						},
-					},
-					{
-						Type: logging.OutputTypeElasticsearch,
-						Name: "es-2",
-						URL:  "https://es-2.svc.messaging.cluster.local:9200",
-						Secret: &logging.OutputSecretSpec{
-							Name: "es-2",
-						},
-					},
-				},
-			},
-			Secrets: map[string]*corev1.Secret{
-				"es-1": {
-					Data: map[string][]byte{
-						"tls.key":       []byte("junk"),
-						"tls.crt":       []byte("junk"),
-						"ca-bundle.crt": []byte("junk"),
-					},
-				},
-				"es-2": {
-					Data: map[string][]byte{
-						"tls.key":       []byte("junk"),
-						"tls.crt":       []byte("junk"),
-						"ca-bundle.crt": []byte("junk"),
-					},
-				},
-			},
-			ExpectedConf: `
-# Logs from containers (including openshift containers)
-[sources.raw_container_logs]
-type = "kubernetes_logs"
-auto_partial_merge = true
-exclude_paths_glob_patterns = ["/var/log/pods/openshift-logging_collector-*/*/*.log", "/var/log/pods/openshift-logging_elasticsearch-*/*/*.log", "/var/log/pods/openshift-logging_kibana-*/*/*.log"]
-
-[sources.raw_journal_logs]
-type = "journald"
-
-# Logs from host audit
-[sources.host_audit_logs]
-type = "file"
-ignore_older_secs = 600
-include = ["/var/log/audit/audit.log"]
-
-# Logs from kubernetes audit
-[sources.k8s_audit_logs]
-type = "file"
-ignore_older_secs = 600
-include = ["/var/log/kube-apiserver/audit.log"]
-
-# Logs from openshift audit
-[sources.openshift_audit_logs]
-type = "file"
-ignore_older_secs = 600
-include = ["/var/log/oauth-apiserver.audit.log"]
-
-[sources.internal_metrics]
-type = "internal_metrics"
-
-[transforms.container_logs]
-type = "remap"
-inputs = ["raw_container_logs"]
-source = """
-  level = "unknown"
-  if match!(.message,r'(Warning|WARN|W[0-9]+|level=warn|Value:warn|"level":"warn")'){
-    level = "warn"
-  } else if match!(.message, r'Info|INFO|I[0-9]+|level=info|Value:info|"level":"info"'){
-    level = "info"
-  } else if match!(.message, r'Error|ERROR|E[0-9]+|level=error|Value:error|"level":"error"'){
-    level = "error"
-  } else if match!(.message, r'Debug|DEBUG|D[0-9]+|level=debug|Value:debug|"level":"debug"'){
-    level = "debug"
-  }
-  .level = level
-  
-  namespace_name = .kubernetes.pod_namespace
-  del(.kubernetes.pod_namespace)
-  .kubernetes.namespace_name = namespace_name
-  
-  del(.file)
-  
-  del(.source_type)
-  
-  del(.stream)
-  
-  del(.kubernetes.pod_ips)
-"""
-
-[transforms.journal_logs]
-type = "remap"
-inputs = ["raw_journal_logs"]
-source = """
-  .
-"""
-
-[transforms.route_container_logs]
-type = "route"
-inputs = ["container_logs"]
-route.app = '!((starts_with!(.kubernetes.namespace_name,"kube")) || (starts_with!(.kubernetes.namespace_name,"openshift")) || (.kubernetes.namespace_name == "default"))'
-route.infra = '(starts_with!(.kubernetes.namespace_name,"kube")) || (starts_with!(.kubernetes.namespace_name,"openshift")) || (.kubernetes.namespace_name == "default")'
-
-# Rename log stream to "application"
-[transforms.application]
-type = "remap"
-inputs = ["route_container_logs.app"]
-source = """
-  .log_type = "application"
-"""
-
-# Rename log stream to "infrastructure"
-[transforms.infrastructure]
-type = "remap"
-inputs = ["route_container_logs.infra","journal_logs"]
-source = """
-  .log_type = "infrastructure"
-"""
-
-# Rename log stream to "audit"
-[transforms.audit]
-type = "remap"
-inputs = ["host_audit_logs","k8s_audit_logs","openshift_audit_logs"]
-source = """
-  .log_type = "audit"
-"""
-
-[transforms.pipeline1]
-type = "remap"
-inputs = ["application","infrastructure"]
-source = """
-  .
-"""
-
-[transforms.pipeline2]
-type = "remap"
-inputs = ["audit"]
-source = """
-  .
-"""
-
-# Adding _id field
-[transforms.es_1_add_es_id]
-type = "remap"
-inputs = ["pipeline1","pipeline2"]
-source = """
-  index = "default"
-  if (.log_type == "application"){
-    index = "app"
-  }
-  if (.log_type == "infrastructure"){
-    index = "infra"
-  }
-  if (.log_type == "audit"){
-    index = "audit"
-  }
-  ."write-index"=index+"-write"
-  ._id = encode_base64(uuid_v4())
-"""
-
-[transforms.es_1_dedot_and_flatten]
-type = "lua"
-inputs = ["es_1_add_es_id"]
-version = "2"
-hooks.process = "process"
-source = """
-    function process(event, emit)
-        if event.log.kubernetes == nil then
-            emit(event)
-            return
-        end
-        dedot(event.log.kubernetes.pod_labels)
-        -- create "flat_labels" key
-        event.log.kubernetes.flat_labels = {}
-        i = 1
-        -- flatten the labels
-        for k,v in pairs(event.log.kubernetes.pod_labels) do
-          event.log.kubernetes.flat_labels[i] = k.."="..v
-          i=i+1
-        end
-        -- delete the "pod_labels" key
-        event.log.kubernetes["pod_labels"] = nil
-        emit(event)
-    end
-
-    function dedot(map)
-        if map == nil then
-            return
-        end
-        local new_map = {}
-        local changed_keys = {}
-        for k, v in pairs(map) do
-            local dedotted = string.gsub(k, "%.", "_")
-            if dedotted ~= k then
-                new_map[dedotted] = v
-                changed_keys[k] = true
-            end
-        end
-        for k in pairs(changed_keys) do
-            map[k] = nil
-        end
-        for k, v in pairs(new_map) do
-            map[k] = v
-        end
-    end
-"""
-
-[sinks.es_1]
-type = "elasticsearch"
-inputs = ["es_1_dedot_and_flatten"]
-endpoint = "https://es-1.svc.messaging.cluster.local:9200"
-index = "{{ write-index }}"
-request.timeout_secs = 2147483648
-bulk_action = "create"
-id_key = "_id"
-
-# TLS Config
-[sinks.es_1.tls]
-key_file = "/var/run/ocp-collector/secrets/es-1/tls.key"
-crt_file = "/var/run/ocp-collector/secrets/es-1/tls.crt"
-
-ca_file = "/var/run/ocp-collector/secrets/es-1/ca-bundle.crt"
-
-# Adding _id field
-[transforms.es_2_add_es_id]
-type = "remap"
-inputs = ["pipeline2"]
-source = """
-  index = "default"
-  if (.log_type == "application"){
-    index = "app"
-  }
-  if (.log_type == "infrastructure"){
-    index = "infra"
-  }
-  if (.log_type == "audit"){
-    index = "audit"
-  }
-  ."write-index"=index+"-write"
-  ._id = encode_base64(uuid_v4())
-"""
-
-[transforms.es_2_dedot_and_flatten]
-type = "lua"
-inputs = ["es_2_add_es_id"]
-version = "2"
-hooks.process = "process"
-source = """
-    function process(event, emit)
-        if event.log.kubernetes == nil then
-            emit(event)
-            return
-        end
-        dedot(event.log.kubernetes.pod_labels)
-        -- create "flat_labels" key
-        event.log.kubernetes.flat_labels = {}
-        i = 1
-        -- flatten the labels
-        for k,v in pairs(event.log.kubernetes.pod_labels) do
-          event.log.kubernetes.flat_labels[i] = k.."="..v
-          i=i+1
-        end
-        -- delete the "pod_labels" key
-        event.log.kubernetes["pod_labels"] = nil
-        emit(event)
-    end
-
-    function dedot(map)
-        if map == nil then
-            return
-        end
-        local new_map = {}
-        local changed_keys = {}
-        for k, v in pairs(map) do
-            local dedotted = string.gsub(k, "%.", "_")
-            if dedotted ~= k then
-                new_map[dedotted] = v
-                changed_keys[k] = true
-            end
-        end
-        for k in pairs(changed_keys) do
-            map[k] = nil
-        end
-        for k, v in pairs(new_map) do
-            map[k] = v
-        end
-    end
-"""
-
-[sinks.es_2]
-type = "elasticsearch"
-inputs = ["es_2_dedot_and_flatten"]
-endpoint = "https://es-2.svc.messaging.cluster.local:9200"
-index = "{{ write-index }}"
-request.timeout_secs = 2147483648
-bulk_action = "create"
-id_key = "_id"
-
-# TLS Config
-[sinks.es_2.tls]
-key_file = "/var/run/ocp-collector/secrets/es-2/tls.key"
-crt_file = "/var/run/ocp-collector/secrets/es-2/tls.crt"
-
-ca_file = "/var/run/ocp-collector/secrets/es-2/ca-bundle.crt"
-
-[sinks.prometheus_output]
-type = "prometheus_exporter"
-inputs = ["internal_metrics"]
-address = "0.0.0.0:24231"
-default_namespace = "collector"
-`,
-		}),
 	)
 	Describe("test helper functions", func() {
 		It("test MakeInputs", func() {
