--- internal/generator/vector/pipelines_to_output.go
@@ -1,16 +1,0 @@
-package vector
-
-import (
-	logging "github.com/openshift/cluster-logging-operator/apis/logging/v1"
-	"github.com/openshift/cluster-logging-operator/internal/generator"
-)
-
-func RouteMap(spec *logging.ClusterLogForwarderSpec, op generator.Options) logging.RouteMap {
-	r := logging.RouteMap{}
-	for _, p := range spec.Pipelines {
-		for _, o := range p.OutputRefs {
-			r.Insert(o, p.Name)
-		}
-	}
-	return r
-}
